# static/openapi.yaml
...
openapi: 3.0.3
info:
  title: OpenAdmin API
  version: 1.0.0
  description: |-
    REST API for OpenAdmin, the administrator-level control panel for OpenPanel server management. Every feature available in OpenAdmin's web UI can also be accessed programmatically.

    All endpoints require a Bearer token obtained from `POST /api/` (username/password login), except the login call itself and the `GET /api/` health check. Tokens expire after 15 minutes. Beyond authentication, most endpoints also require the administrator role -- see each operation's **Access** note for its exact requirement (some allow a reseller who owns the target account, and a handful of the most dangerous server-control actions require the Super Admin role specifically).

    The whole API is also gated behind [PANEL] api=on and a configured license key -- when either is off, every /api/* route (including this one) responds as if it doesn't exist.

    Success responses are JSON objects (`200`) unless noted otherwise. Error responses always include an `error` field and use standard HTTP status codes (400, 401, 403, 404, 409, 500).
  contact:
    name: OpenPanel
    url: https://openpanel.com
  license:
    name: OpenPanel EULA
    url: https://openpanel.com/LICENSE
servers:
- url: /
  description: This OpenAdmin installation
...
paths:
...
  /api/settings/locales:
    get:
      summary: Lists translation locales available on GitHub, marking which are installed locally and
        which one is the current default.
      description: '**Access:** Administrator or user role (the reseller role is blocked).'
      operationId: get_settings_locales
      tags:
      - Settings - Locales
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties:
                  default_locale:
                    type: string
                    example: en
                  translations:
                    type: array
                    items:
                      type: object
                      properties:
                        locale:
                          type: string
                          example: en
                        provider:
                          type: string
                          example: OpenPanel
                        path:
                          type: string
                          example: /etc/openpanel/openadmin/translations/en
                        installed:
                          type: boolean
                          example: true
                        is_default:
                          type: boolean
                          example: true
                        update_available:
                          type: boolean
                          description: True when the installed file differs from the latest version
                            on GitHub.
                          example: false
                      example:
                        locale: en
                        provider: OpenPanel
                        path: /etc/openpanel/openadmin/translations/en
                        installed: true
                        is_default: true
                    example:
                    - locale: en
                      provider: OpenPanel
                      path: /etc/openpanel/openadmin/translations/en
                      installed: true
                      is_default: true
                    - locale: de-DE
                      provider: Community
                      path: ''
                      installed: false
                      is_default: false
                example:
                  default_locale: en
                  translations:
                  - locale: en
                    provider: OpenPanel
                    path: /etc/openpanel/openadmin/translations/en
                    installed: true
                    is_default: true
                  - locale: de-DE
                    provider: Community
                    path: ''
                    installed: false
                    is_default: false
        default:
          $ref: '#/components/responses/Error'
      security:
      - bearerAuth: []
    post:
      summary: Installs a locale ('locale' field), updates an installed one ('update' field), deletes
        one ('delete' field), or sets the panel's default locale ('default' field); exactly one field is
        expected per call. Pass 'all' to 'locale' to install every available locale, or to 'update' to
        update every locale that has an update available.
      description: '**Access:** Administrator or user role (the reseller role is blocked).


        With `{"locale": "all"}` or any `update` value the response also carries `installed` and `failed`
        arrays, and `success` is false if any locale failed to download. The default locale can''t be
        deleted.'
      operationId: post_settings_locales
      tags:
      - Settings - Locales
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                locale:
                  type: string
                  example: de-DE
                update:
                  type: string
                  example: ''
                delete:
                  type: string
                  example: ''
                default:
                  type: string
                  example: ''
              example:
                locale: de-DE
                default: ''
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Locale 'de-DE' installed successfully.
                  installed:
                    type: array
                    description: Only returned when locale is 'all' or for update. Locales that were installed
                      or updated.
                    items:
                      type: string
                    example:
                    - de-de
                    - sr-rs
                  failed:
                    type: array
                    description: Only returned when locale is 'all' or for update.
                    items:
                      type: string
                    example: []
                example:
                  success: true
                  message: Locale 'de-DE' installed successfully.
        default:
          $ref: '#/components/responses/Error'
      security:
      - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Obtain a token via `POST /api/` (username/password login). Tokens expire after 15 minutes.
  responses:
    Error:
      description: An error occurred. See the `error` field for details.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 'Forbidden: administrator role required'
            required:
            - error
security:
- bearerAuth: []